Use GoComics's timezone
This change sets all datetime objects to GoComics's timezone, which I deduced through my own testing is UTC-5. This makes requests more accurate. For example, if you request a comic for dt.datetime.now() in a timezone that is currently a day ahead of GoComics, the code would throw an error under the old behavior, since it would be asking for a comic that hasn't been published yet (GoComics always publishes at midnight). Under the new behavior the code will always request the comic for the day it is _for GoComics_.